requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
pypdf2
pdfplumber
fastapi
uvicorn
sse_starlette
SpeechRecognition
pyaudio
gtts
pygame
faiss-cpu
mysql-connector-python
translate
langchain_groq
chromadb